Added attribution for music

+++ Application/src/application/app/CreatePreview.java
@@ -40,9 +40,9 @@ public class CreatePreview extends Task<Void> {
 	        InputStream stdout = process.getInputStream();
 	        BufferedReader stdoutBuffered = new BufferedReader(new InputStreamReader(stdout));
 	        String line = null;
-	        int percentage;
+	        double percentage;
 	        while ((line = stdoutBuffered.readLine()) != null ) {
-	        	percentage = Integer.parseInt(line);
+	        	percentage = Double.parseDouble(line);
 	        	updateProgress(percentage,100); //ProgressBar listens to this line
 	        }
 	        stdoutBuffered.close();
